+++ src/niobot/client.py
@@ -29,7 +29,7 @@ from .exceptions import (
     MessageException,
     NioBotException,
 )
-from .utils import Typing, deprecated, force_await, run_blocking
+from .utils import Mentions, Typing, deprecated, force_await, run_blocking
 from .utils.help_command import default_help_command
 
 try:
@@ -220,6 +220,8 @@ class NioBot(nio.AsyncClient):
             self.__key_import = pathlib.Path(keys_path), keys_password
         else:
             self.__key_import = None
+
+        self.server_info: typing.Optional[dict] = None
         self.add_command(help_cmd)
 
     async def sync(self, *args, **kwargs) -> U[nio.SyncResponse, nio.SyncError]:
@@ -868,11 +870,12 @@ class NioBot(nio.AsyncClient):
         room: U[nio.MatrixRoom, nio.MatrixUser, str],
         content: typing.Optional[str] = None,
         file: typing.Optional[BaseAttachment] = None,
-        reply_to: typing.Optional[U[nio.RoomMessageText, str]] = None,
+        reply_to: typing.Optional[U[nio.RoomMessage, str]] = None,
         message_type: typing.Optional[str] = None,
         *,
         content_type: typing.Literal["plain", "markdown", "html", "html.raw"] = "markdown",
         override: typing.Optional[dict] = None,
+        mentions: typing.Optional[Mentions] = None,
     ) -> nio.RoomSendResponse:
         """
         Sends a message. Doesn't get any more simple than this.
@@ -906,6 +909,7 @@ class NioBot(nio.AsyncClient):
         :param override: A dictionary containing additional properties to pass to the body.
         Overrides existing properties.
         :param content_type: The type of content to send. Defaults to "markdown".
+        :param mentions: Intentional mentions to send with the message.
         :return: The response from the server.
         :raises MessageException: If the message fails to send, or if the file fails to upload.
         :raises ValueError: You specified neither file nor content.
@@ -983,10 +987,19 @@ class NioBot(nio.AsyncClient):
 
         if reply_to:
             body["m.relates_to"] = {"m.in_reply_to": {"event_id": self._get_id(reply_to)}}
+        if mentions:
+            body.update(mentions.as_body())
 
         if override:
             body.update(override)
-        async with Typing(self, self._get_id(room)):
+        try:
+            async with Typing(self, self._get_id(room)):
+                response = await self.room_send(
+                    self._get_id(room),
+                    "m.room.message",
+                    body,
+                )
+        except RuntimeError:  # already typing
             response = await self.room_send(
                 self._get_id(room),
                 "m.room.message",
@@ -1004,6 +1017,7 @@ class NioBot(nio.AsyncClient):
         *,
         message_type: typing.Optional[str] = None,
         content_type: typing.Literal["plain", "markdown", "html", "html.raw"] = "markdown",
+        mentions: typing.Optional[Mentions] = None,
         override: typing.Optional[dict] = None,
     ) -> nio.RoomSendResponse:
         """
@@ -1061,6 +1075,8 @@ class NioBot(nio.AsyncClient):
         elif content_type == "html.raw":
             body["formatted_body"] = content
             body["format"] = "org.matrix.custom.html"
+        if mentions:
+            body.update(mentions.as_body())
         if override:
             body.update(override)
         async with Typing(self, room):
@@ -1181,6 +1197,13 @@ class NioBot(nio.AsyncClient):
                 self.log.critical("Failed to upload encryption keys. Encryption may not work. Error: %r", response)
             else:
                 self.log.info("Uploaded encryption keys.")
+        self.log.info("Fetching server details...")
+        response = await self.send("GET", "/_matrix/client/versions")
+        if response.status != 200:
+            self.log.warning("Failed to fetch server details. Status: %d", response.status)
+        else:
+            self.server_info = await response.json()
+            self.log.debug("Server details: %r", self.server_info)
         self.log.info("Performing first sync...")
         result = await self.sync(timeout=30000, full_state=True, set_presence="unavailable")
         if not isinstance(result, nio.SyncResponse):