+++ src/niobot/utils/__init__.py
@@ -2,6 +2,7 @@ from .checks import *
 from .federation import *
 from .help_command import *
 from .lib import *
+from .mentions import Mentions as Mentions
 from .parsers import *
 from .string_view import *
 from .typing import *