+++ src/niobot/utils/help_command.py
@@ -1,7 +1,6 @@
 import re
 import textwrap
 import typing
-import warnings
 
 if typing.TYPE_CHECKING:
     from ..commands import Command
@@ -13,7 +12,6 @@ __all__ = (
     "format_command_line",
     "get_short_description",
     "get_long_description",
-    "help_command_callback",
     "default_help_command",
     "clean_output",
 )
@@ -140,12 +138,3 @@ async def default_help_command(ctx: "Context"):
         description = get_long_description(command)
         lines = ["* {}:".format(display), *description.splitlines()]
         await ctx.respond(clean_output("\n".join(lines)))
-
-
-def help_command_callback(ctx: "Context"):
-    """Default help command callback"""
-    warnings.warn(
-        "help_command_callback is deprecated and will be removed in v1.2.0, please use default_help_command instead",
-        DeprecationWarning,
-    )
-    return default_help_command(ctx)